backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document processing
python-docx==0.8.11
pypdf==3.17.4
jinja2==3.1.2
weasyprint==61.2
pydyf<0.11.0

# AI and NLP
google-cloud-aiplatform==1.40.0
google-cloud-secret-manager==2.18.1
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Networking and security
httpx==0.25.2
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Web scraping (for LinkedIn import)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0